+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/resolver/PhysicalDataSourceConstraint.java
@@ -47,7 +47,6 @@ public class PhysicalDataSourceConstraint extends DataSourceConstraint {
     ) {
         super(dataSourceConstraint);
         this.allColumnPhysicalNames = allColumnPhysicalNames;
-
     }
 
     /**